go-logging: provide right pc to runtime.FuncForPC
We want a PC in the call instruction, not the PC of the instruction
just after the call instruction.

(A more robust fix would be to move to runtime.CallersFrames, as
runtime.Callers+runtime.FuncForPC is deprecated.)

This updates go-logging so it works with the mid-stack inlining coming
with go 1.12.